Desktop/AE-REAL-ESTATE/agency-ai-whitelabel/agents-python/requirements.txt
fastapi
uvicorn
pydantic
requests
crewai
langchain-openai
python-dotenv
